feat: 1# 添加读写数据的测试，并修改错误; 2# DataItem的count使用位进行计数的; 3# 写入数据的返回数据没有相应的类型和长度;

+++ src/main/java/com/github/xingshuangs/iot/protocol/s7/model/COTPConnection.java
@@ -111,10 +111,10 @@ public class COTPConnection extends COTP implements IByteArray {
     @Override
     public byte[] toByteArray() {
         byte[] res = new byte[BYTE_LENGTH];
-        byte[] destRefBytes = ShortUtil.toByteArray((short) this.destinationReference);
-        byte[] srcRefBytes = ShortUtil.toByteArray((short) this.sourceReference);
-        byte[] srcTsapBytes = ShortUtil.toByteArray((short) this.sourceTsap);
-        byte[] destTsapBytes = ShortUtil.toByteArray((short) this.destinationTsap);
+        byte[] destRefBytes = ShortUtil.toByteArray(this.destinationReference);
+        byte[] srcRefBytes = ShortUtil.toByteArray(this.sourceReference);
+        byte[] srcTsapBytes = ShortUtil.toByteArray(this.sourceTsap);
+        byte[] destTsapBytes = ShortUtil.toByteArray(this.destinationTsap);
 
         res[0] = ByteUtil.toByte(this.length);
         res[1] = this.pduType.getCode();
