+++ src/test/java/com/github/xingshuangs/iot/utils/ShortUtilTest.java
@@ -9,14 +9,14 @@ public class ShortUtilTest {
 
     @Test
     public void toByteArray() {
-        byte[] actual = ShortUtil.toByteArray((short) 24565);
+        byte[] actual = ShortUtil.toByteArray(24565);
         byte[] expect = new byte[]{(byte) 0x5F, (byte) 0xF5};
         assertArrayEquals(expect, actual);
     }
 
     @Test
     public void toByteArray1() {
-        byte[] actual = ShortUtil.toByteArray((short) 24565, true);
+        byte[] actual = ShortUtil.toByteArray(24565, true);
         byte[] expect = new byte[]{(byte) 0xF5, (byte) 0x5F};
         assertArrayEquals(expect, actual);
     }
